Fix the crash bug when removing a printer, and added a printer delegate
svn path=/trunk/playground/base/print-manager/; revision=1102946

+++ printer-manager-kcm/PrintKCM.cpp
@@ -21,8 +21,10 @@
 #include "PrintKCM.h"
 
 #include "PrinterModel.h"
+#include "PrinterDelegate.h"
 #include "PrinterDescription.h"
 
+#include <KMessageBox>
 #include <KGenericFactory>
 #include <KAboutData>
 #include <KIcon>
@@ -53,37 +55,47 @@ PrintKCM::PrintKCM(QWidget *parent, const QVariantList &args)
     removePB->setIcon(KIcon("list-remove"));
     preferencesPB->setIcon(KIcon("configure"));
 
-    // setup the timer that updates the UIs
-    QTimer *updateT = new QTimer(this);
-    updateT->setInterval(1000);
-    updateT->start();
-    connect(updateT, SIGNAL(timeout()),
-            this, SLOT(update()));
-
     QCups::initialize();
     m_model = new PrinterModel(winId(), this);
     printersTV->setModel(m_model);
+    printersTV->setItemDelegate(new PrinterDelegate(this));
     connect(printersTV->selectionModel(), SIGNAL(selectionChanged(const QItemSelection &, const QItemSelection &)),
             this, SLOT(update()));
+    connect(printersTV->model(), SIGNAL(dataChanged(const QModelIndex &, const QModelIndex &)),
+            this, SLOT(update()));
 
-    m_printerDesc = new PrinterDescription;
+    // setup the timer that updates the UIs
+    m_updateT = new QTimer(this);
+    m_updateT->setInterval(1000);
+    m_updateT->start();
+    connect(m_updateT, SIGNAL(timeout()),
+            m_model, SLOT(update()));
+
+    // Create the PrinterDescription before we try to select a printer
+    m_printerDesc = new PrinterDescription(this);
     m_printerDesc->hide();
+
+    // select the first printer if there are printers
+    if (m_model->rowCount() > 0) {
+        printersTV->selectionModel()->select(m_model->index(0, 0), QItemSelectionModel::Select);
+    }
 }
 
 PrintKCM::~PrintKCM()
 {
-    delete m_printerDesc;
+    m_updateT->stop();
 }
 
 void PrintKCM::update()
 {
-    m_model->update();
     QItemSelection selection;
     // we need to map the selection to source to get the real indexes
     selection = printersTV->selectionModel()->selection();
     // enable or disable the job action buttons if something is selected
     if (!selection.indexes().isEmpty()) {
         if (scrollArea->widget() != m_printerDesc) {
+            // always take the widget before setting a new one otherwise it will be deleted
+            scrollArea->takeWidget();
             scrollArea->setWidget(m_printerDesc);
         }
         removePB->setEnabled(true);
@@ -91,11 +103,13 @@ void PrintKCM::update()
         m_printerDesc->setDestName(index.data(PrinterModel::DestName).toString());
         m_printerDesc->setLocation(index.data(PrinterModel::DestLocation).toString());
         m_printerDesc->setStatus(index.data(PrinterModel::DestStatus).toString());
-        m_printerDesc->setDescription(index.data(PrinterModel::DestDescription).toString());
+        m_printerDesc->setDescription(index.data(Qt::DisplayRole).toString());
         m_printerDesc->setKind(index.data(PrinterModel::DestKind).toString());
         m_printerDesc->setIsShared(index.data(PrinterModel::DestIsShared).toBool());
         m_printerDesc->setIsDefault(index.data(PrinterModel::DestIsDefault).toBool());
     } else if (!noPrinterL->isVisible()) {
+        // always take the widget before setting a new one otherwise it will be deleted
+        scrollArea->takeWidget();
         scrollArea->setWidget(noPrinterL);
         removePB->setEnabled(false);
     }
@@ -109,6 +123,13 @@ void PrintKCM::on_removePB_clicked()
     // enable or disable the job action buttons if something is selected
     if (!selection.indexes().isEmpty()) {
         QModelIndex index = selection.indexes().at(0);
-        QCups::deletePrinter(index.data(PrinterModel::DestName).toString());
+        int resp;
+        resp = KMessageBox::questionYesNo(this,
+                                          i18n("Are you sure you want to remove the printer '%1'?",
+                                               index.data(Qt::DisplayRole).toString()),
+                                          i18n("Remove printer"));
+        if (resp == KMessageBox::Yes) {
+            QCups::deletePrinter(index.data(PrinterModel::DestName).toString());
+        }
     }
 }
